- Fix CI errors.

+++ shared/properties-view/double-slider.cpp
@@ -27,5 +27,5 @@ void DoubleSlider::setDoubleConstraints(double newMin, double newMax,
 
 void DoubleSlider::setDoubleVal(double val)
 {
-	setValue(lround((val - minVal) / minStep));
+	setValue((int)lround((val - minVal) / minStep));
 }
